requirements.txt
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.0
Django==1.9.7
django-rest-framework-mongoengine==3.3.0
django-rest-swagger==0.3.8
django-test-addons==1.0
djangorestframework==3.3.3
docker-compose==1.5.0
docker-py==1.8.1
dockerpty==0.3.4
docopt==0.6.2
ecdsa==0.13
Fabric3==1.11.1.post1
jsonschema==2.5.1
mongodbforms==0.3
mongoengine==0.9.0
paramiko==1.17.1
pycrypto==2.6.1
pymongo==2.8.1
python-dateutil==2.5.3
python-mimeparse==1.5.2
PyYAML==3.11
requests==2.7.0
six==1.10.0
texttable==0.8.4
websocket-client==0.37.0
